agent_check v0.2.1 AgentCheck

Implementation of the Agent Check protocol for your Elixir/Phoenix app. Also see - https://cbonte.github.io/haproxy-dconv/1.8/configuration.html#5.2-agent-check

It allows for easy rolling restarts and dynamic backpressure to your loadbalancer.

Functions

accept(port)

Starts accepting connections on the given port.

format_haproxy_state(stats_struct)

Reformat the stats struct into a haproxy state format.

handle_command(command)

loop_acceptor(socket)

Wait for incomming socket connection (blocking) and spawn of a Task when one comes in.

read_line(socket)

serve(socket)